moving copy to RPC
